Rename python parts
Avoid leaking Axis class names

Drop underscores infront of _any
