+++ boost_histogram/__init__.py
@@ -2,10 +2,10 @@ from __future__ import absolute_import, division, print_function
 
 del absolute_import, division, print_function
 
-from ._hist import histogram
+from ._internal.hist import histogram
 
 from . import axis, storage, accumulators, algorithm, numpy
 
-from .uhi import loc, rebin, project, underflow, overflow
+from .tag import loc, rebin, project, underflow, overflow
 
 from .version import __version__
