+++ boost_histogram/_internal/axistuple.py
