+++ boost_histogram/tag.py
